Move some functions to a file not in the compiler
