+++ src/printing.cc
@@ -39,11 +39,6 @@ void print_name_console(String* string) {
   print_name(&p, string);
 }
 
-void print_heap_console(ObjectHeap* heap, const char* title) {
-  ConsolePrinter p(null);
-  print_heap(&p, heap, title);
-}
-
 #define BYTECODE_PRINT(name, length, format, print) print,
 static const char* opcode_print[] { BYTECODES(BYTECODE_PRINT) "Illegal" };
 #undef BYTECODE_PRINT
@@ -453,13 +448,6 @@ void print_object(Printer* printer, Object* object) {
 void print_object_short(Printer* printer, Object* object, bool is_top_level) {
   ShortPrintVisitor p(printer, is_top_level);
   p.accept(object);
-}
-
-void print_heap(Printer* printer, ObjectHeap* heap, const char* title) {
-  printer->printf("%s:\n", title);
-  heap->do_objects([&] (HeapObject* object) -> void {
-    print_object(printer, object);
-  });
 }
 
 void Printer::print_buffer(const uint8_t* s, int len) {
