+++ src/vm.cc
@@ -21,6 +21,7 @@
 #include "objects_inline.h"
 #include "os.h"
 #include "primitive.h"
+#include "printing.h"
 #include "resource.h"
 #include "scheduler.h"
 #include "vm.h"
@@ -58,5 +59,21 @@ VM::~VM() {
 }
 
 VM* VM::_current = null;
+
+#ifdef DEBUG
+
+void print_heap_console(ObjectHeap* heap, const char* title) {
+  ConsolePrinter p(null);
+  print_heap(&p, heap, title);
+}
+
+void print_heap(Printer* printer, ObjectHeap* heap, const char* title) {
+  printer->printf("%s:\n", title);
+  heap->do_objects([&] (HeapObject* object) -> void {
+    print_object(printer, object);
+  });
+}
+
+#endif
 
 } // namespace toit
